requirements.txt
anyio==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
fastapi==0.75.0
idna==3.3
pydantic==1.9.0
Rastreio-Correios==0.1.4
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.8
